task

+++ 0x08-recursion/holberton.h
@@ -10,7 +10,7 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int is_prime_number(int n);
-int foo_recursion(int n, int i);
+int foo_recursion(int n, long int i);
 int _sqrt_recursion(int n);
 int _sqrt_recursive(int n, int i);
